feat: Add user_id validation to attachment detach method
The `detach` method in the `Attachment` struct now accepts an additional `user_id` parameter to validate if the user has permission to delete the attachment. If the `user_id` does not match the `attached.user_id`, an error is returned with the message "You do not have permission to delete this attachment". This change ensures that only the authorized user can delete attachments.

+++ create-rust-app/src/storage/attachment.rs
@@ -237,10 +237,14 @@ impl Attachment {
     /// # Errors
     /// * Diesel error
     #[cfg(feature = "backend_actix-web")]
-    pub async fn detach(db: &mut Connection, storage: &Storage, item_id: ID) -> Result<(), String> {
+    pub async fn detach(db: &mut Connection, storage: &Storage, item_id: ID, user_id: ID) -> Result<(), String> {
         let attached = Self::find_by_id(db, item_id).map_err(|_| "Could not load attachment")?;
         let blob = AttachmentBlob::find_by_id(db, attached.blob_id)
             .map_err(|_| "Could not load attachment blob")?;
+
+        if user_id != attached.user_id {
+            return Err("You do not have permission to delete this attachment".to_string());
+        }
 
         let delete_result = storage.delete(blob.key.clone()).await;
